requirements.txt
# Core Dependencies
fastapi>=0.115.11
uvicorn>=0.34.0
python-dotenv>=1.0.1
pydantic>=2.10.6

# LangChain & AI
langchain>=0.0.300
langchain-core>=0.3.45
langchain-community>=0.2.17
langchain-openai>=0.3.9
langchain-chroma>=0.2.2
langchain-milvus>=0.1.8
langchain-ollama>=0.2.3
openai
tiktoken>=0.9.0
pymilvus<=3.0.0
selenium>=4.21.0
webdriver-manager==4.0.1  # Tự động quản lý trình điều khiển

# Vector Database & Search
chromadb>=0.6.3
sqlalchemy>=2.0.39
rank-bm25>=0.2.2

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
python-multipart>=0.0.20
beautifulsoup4>=4.13.3

# UI/Frontend
streamlit>=1.43.2

# Utils
requests>=2.32.3
tqdm>=4.67.1
PyYAML>=6.0.2

#gemini 
langchain-google-genai>=2.1.0
langchain-google-vertexai>=2.0.15
langchain-huggingface<= 0.1.2
vncorenlp<=1.0.3
langchain-experimental<=0.3.4
sentence-transformers<=3.4.1

astrapy>=2.0,<3.0
